+++ active_adaptation/envs/mdp/observations/extero/height_map.py
@@ -46,8 +46,10 @@ class external_forces(Observation):
         )
 
     def update(self):
-        forces_b = self.asset._external_force_b[:, self.body_ids]
-        forces_b /= self.denom
+        forces_b = self.asset.permanent_wrench_composer.composed_force_as_torch[
+            :, self.body_ids
+        ]
+        forces_b = forces_b / self.denom
         self.forces_b = forces_b
 
     def compute(self) -> torch.Tensor:
@@ -79,7 +81,9 @@ class external_torques(Observation):
         )
 
     def update(self):
-        torques_b = self.asset._external_torque_b[:, self.body_ids]
+        torques_b = self.asset.permanent_wrench_composer.composed_torque_as_torch[
+            :, self.body_ids
+        ]
         torques_b = torques_b / self.denom
         self.torques_b = torques_b