+++ projects/facet/facet_commands/impedance_manip.py
@@ -379,17 +379,16 @@ class ImpedanceCommandManager(Command):
             raise ValueError(f"Invalid input dimension: {x.ndim}")
         return quat_rotate(quat, x)
 
-    def step(self, substep: int):
-        # self.asset._external_force_b[:, self.eef_body_id] += quat_rotate_inverse(
-        #     self.asset.data.body_quat_w[:, self.eef_body_id],
-        #     self.get_eef_force(self.asset.data.body_pos_w[:, self.eef_body_id], self.asset.data.body_lin_vel_w[:, self.eef_body_id])
-        # )
+    def pre_step(self, substep: int):
         force_base_w = self.force_impulse.get_force(None, None)
-        self.asset._external_force_b[:, 0] += quat_rotate_inverse(
+        force_b = quat_rotate_inverse(
             self.asset.data.root_link_quat_w,
-            force_base_w
+            force_base_w,
+        ).unsqueeze(1)
+        self.asset.permanent_wrench_composer.add_forces_and_torques(
+            forces=force_b,
+            body_ids=[0],
         )
-        self.asset.has_external_wrench = True
 
     def reset(self, env_ids: torch.Tensor, tensordict: TensorDictBase):
         root_link_pos = self.asset.data.root_link_pos_w[env_ids]